refactor(gabinet): co-locate ChangePasswordDialog state in AccountTabContent (closes #4458)
Move changePasswordOpen, newPassword, confirmPassword, changePasswordSubmitting,
and changePasswordError state — plus validation and submit logic — out of
EmployeeDetail and into AccountTabContent. EmployeeDetail now passes the
changeEmployeePassword mutation as a simple (newPassword: string) => Promise<void>
callback, removing the state leak between the route component and the account tab.
